Start to remove redis dependency by using APC as a cache.
Create generic cache class to enable different backends (APC, Redis, File, ...). We still have the issue that we cannot use APC for everything because the cache is not shared between Apache PHP and cli PHP, which is used by phpcron. A solution could be to run phpcron via apache2 requests.